fix(app_configuration): decompose app_configuration at enviroment_config and vue_functionallity; make right order of resource loading;
