+++ .vscode/settings.json
@@ -1,61 +1,5 @@
 {
-    // "editor.defaultFormatter": "dbaeumer.vscode-eslint",
-    // "editor.quickSuggestions": {
-    //     "strings": true
-    // },
-    // "[javascript]": {
-    //     "editor.defaultFormatter": "vscode.typescript-language-features"
-    // },
-    // "[typescript]": {
-    //     "editor.defaultFormatter": "dbaeumer.vscode-eslint"
-    // },
-    // "[vue]": {
-    //     "editor.defaultFormatter": "dbaeumer.vscode-eslint"
-    // },
-    // "[scss]": {
-    //     "editor.defaultFormatter": "stylelint.vscode-stylelint"
-    // },
-    // "[css]": {
-    //     "editor.defaultFormatter": "stylelint.vscode-stylelint"
-    // },
-    // "editor.codeActionsOnSave": {
-    //     "source.fixAll.eslint": true,
-    //     "source.fixAll.stylelint": true
-    // },
-    // "eslint.alwaysShowStatus": true,
-    // "eslint.options": {
-    //     "extensions": [
-    //         ".html",
-    //         ".js",
-    //         ".ts",
-    //         ".vue"
-    //     ]
-    // },
-    // "stylelint.enable": true,
     "stylelint.validate": [ "css", "scss", "vue" ],
-    // "css.validate": true,
-    // "scss.validate": true,
-    // "html.format.enable": false,
-    // "emmet.triggerExpansionOnTab": true,
-    // "emmet.includeLanguages": {
-    //     "vue-html": "html"
-    // },
-    // "files.exclude": {
-    //     "**/*.log": true,
-    //     "**/*.log*": true,
-    //     "**/dist": true,
-    // },
-    // "files.associations": {
-    //     ".babelrc": "jsonc",
-    //     ".eslintrc": "jsonc",
-    //     ".markdownlintrc": "jsonc",
-    //     "*.config.js": "javascript",
-    //     "*.spec.js": "javascript",
-    //     "*.vue": "vue"
-    // },
-    // "files.eol": "\n",
-    // "files.insertFinalNewline": true,
-    // "files.trimFinalNewlines": true,
-    // "files.trimTrailingWhitespace": true,
-    // "tailwindCSS.emmetCompletions": true,
+    "css.validate": false,
+    "scss.validate": false,
 }
