+++ src/configuration/environment/_utils/registrateGlobalSafely.js
