+++ src/configuration/environment/index.js
@@ -1,0 +1,5 @@
+import { registrateGlobals } from './registrateGlobals.js'
+
+export {
+  registrateGlobals,
+}
