+++ src/configuration/vue/index.js
