--- src/helpers/app_configuration/index.js
@@ -1,7 +1,0 @@
-import { registrateGlobals } from './registrateGlobals.js'
-import { registrateVueFunctionality } from './registrateVueFunctionality.js'
-
-export {
-  registrateGlobals,
-  registrateVueFunctionality,
-}
