+++ src/main.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-expressions */
-import * as envConfiguration from '@/helpers/app_configuration'
+import * as envConfiguration from '@/configuration/environment'
 
 envConfiguration.registrateGlobals()
 
 import( /* webpackPreload: true */ '@services' )
-import( /* webpackPreload: true */'@scss/_predefined.scss' )
+import( /* webpackPreload: true */ '@scss/_predefined.scss' )
 
 import( './app' )
